Refactor book details screen and related widgets
This commit refactors the book details screen and related widgets to use the `BookModel` class instead of hardcoding values. It also adds the `CachedNetworkImage` widget to display book cover images and updates the `CustomBookDetailsRating` widget to display the actual average rating and rating count. Additionally, the `SimiliarBooksList` widget now displays the actual list of books instead of a hardcoded number.